added requirements file

+++ teammembers/views.py
@@ -6,6 +6,11 @@ from .forms import TeamMemberModelForm
 # Create your views here.
 
 class TeamMemberEditView(View):
+	"""
+	View for editing
+	If save button is pressed, saves edited teammember and redirects to list page
+	Delete href on page that can redirect to a delete page
+	"""
 	template_name = "teammembers/teammember_edit.html"
 	def get_object(self):
 		id = self.kwargs.get('id')
@@ -15,6 +20,7 @@ class TeamMemberEditView(View):
 		return obj
 
 	def get(self, request, id=None, *args, **kwargs):
+		# GET method
 		context = {}
 		obj = self.get_object()
 		if obj is not None:
@@ -24,49 +30,34 @@ class TeamMemberEditView(View):
 		return render(request, self.template_name, context)
 
 	def post(self, request, id=None, *args, **kwargs):
+		# POST method
 		context = {}
 		obj = self.get_object()
 		if obj is not None:
+			# form with POST method and individual teammember
 			form = TeamMemberModelForm(request.POST, instance=obj)
 			if request.method == "POST":
-				if form.is_valid():
-					form.save()
-				context['object'] = obj
-				context['form'] = form
+				if 'save_button' in request.POST:
+					# if save submit button pushed
+					if form.is_valid():
+						form.save()
+					context['object'] = obj
+					context['form'] = form
+				if 'del_button' in request.POST:
+					# if delete button pushed
+					obj.delete()
+					context['object'] = obj
 				return redirect('/teammembers/')
 		return render(request, self.template_name, context) 
 
 
-
-class TeamMemberDeleteView(View):
-	template_name = "teammembers/teammember_delete.html"
-	def get_object(self):
-		id = self.kwargs.get('id')
-		if id is not None:
-			obj = get_object_or_404(TeamMember, id=id)
-		return obj
-
-	def get(self, request, id=None, *args, **kwargs):
-		context = {}
-		obj = self.get_object()
-		if obj is not None:
-			context['object'] = obj
-		return render(request, self.template_name, context)
-
-	def post(self, request, id=None, *args, **kwargs):
-		context = {}
-		obj = self.get_object()
-		if obj is not None:
-			obj.delete()
-			context['object'] = obj
-			return redirect('/teammembers/')
-		return render(request, self.template_name, context) 
-
-
-
 class TeamMemberUpdateView(View):
+	"""
+	Original view for changing existing data for teammembers, could not delete
+	"""
 	template_name = "teammembers/teammember_update.html"
 	def get_object(self):
+		# GET object method
 		id = self.kwargs.get('id')
 		obj = None
 		if id is not None:
@@ -74,6 +65,7 @@ class TeamMemberUpdateView(View):
 		return obj
 
 	def get(self, request, id=None, *args, **kwargs):
+		# GET method
 		context = {}
 		obj = self.get_object()
 		if obj is not None:
@@ -83,52 +75,100 @@ class TeamMemberUpdateView(View):
 		return render(request, self.template_name, context)
 
 	def post(self, request, id=None, *args, **kwargs):
+		# POST method
 		context = {}
 		obj = self.get_object()
 		if obj is not None:
+			# form with POST method passed in and individual teammember
 			form = TeamMemberModelForm(request.POST, instance=obj)
 			if form.is_valid():
+				# saves form
 				form.save()
 			context['object'] = obj
 			context['form'] = form
 		return render(request, self.template_name, context) 
 
 
+
+class TeamMemberDeleteView(View):
+	"""
+	Original view for deleting, now done in edit view
+	"""
+	template_name = "teammembers/teammember_delete.html"
+	def get_object(self):
+		# GET object method
+		id = self.kwargs.get('id')
+		if id is not None:
+			obj = get_object_or_404(TeamMember, id=id)
+		return obj
+
+	def get(self, request, id=None, *args, **kwargs):
+		# GET method
+		context = {}
+		obj = self.get_object()
+		if obj is not None:
+			context['object'] = obj
+		return render(request, self.template_name, context)
+
+	def post(self, request, id=None, *args, **kwargs):
+		# POST method
+		context = {}
+		obj = self.get_object()
+		if obj is not None:
+			obj.delete()
+			context['object'] = obj
+			# redirects to list page
+			return redirect('/teammembers/')
+		return render(request, self.template_name, context) 
+
+
 class TeamMemberCreateView(View):
 	template_name = "teammembers/teammember_add.html"
 	def get(self, request, *args, **kwargs):
+		#GET method
+		# empty context for form
 		form = TeamMemberModelForm()
 		context = {"form": form}
-		#print(self)
 		return render(request, self.template_name, context)
 
 	def post(self, request, *args, **kwargs):
+		#POST method
+		# Model form with POST method
 		form = TeamMemberModelForm(request.POST)
 		if form.is_valid():
 			form.save()
 		#form = TeamMemberModelForm()
 		if request.method == "POST":
+			# redirects back to list page
 			return redirect('../')
 		context = {"form": form}
 		return render(request, self.template_name, context)
 
 
 class TeamMemberListView(View):
+	"""
+	View for printing all of the teammembers
+	"""
 	template_name = "teammembers/teammember_list.html"
 	queryset = TeamMember.objects.all()
-
+	"""
 	def get_queryset(self):
 		return self.queryset
+	"""
 	def get(self, request, *args, **kwargs):
 		#queryset = self.get_queryset
 		#print(queryset)
-
 		#context = {'object_list': self.get_queryset()}
+
+		# queryset of all teammembers is context
 		queryset = TeamMember.objects.all()
 		context = {'object_list': queryset}
 		return render(request, self.template_name, context)
 
 class TeamMemberView(View):
+	"""
+	View of individual teammember
+	"""
 	template_name = "teammembers/teammember_detail.html"
 	def get(self, request, id=None, *args, **kwargs):
 		context = {}
